Changeset indexes by node id instead of path
Main reason to implement this at this point is to facilitate using a KV
store for the manifests. The key in that case will contain the node id,
and having the changeset indexed by node id will help.

As a side-effect this will make it easier to implement move/rename
functionality.