Cache assets independently and by count,  not by mem
This is acknowledging that in the new in-memory datastructure model we
cannot easily know memory usage for the assets, and that we probably
don't want anyway.

Now the caching storage caches assets independently by type and limits
the cache size by count of cached assets.

We introduce a function Dataset:

```
  pub fn add_in_mem_asset_caching(
      storage: Arc<dyn Storage + Send + Sync>,
  ) -> Arc<dyn Storage + Send + Sync> {
```

That provides the default, recommended level of caching to a `Storage`.

Once we have more experience with performance, we will be able to offer
more powerful caching and configuration.