Make local filesystem Storage implementation work
The issue was that object_store doesn't return lists in sorted order. So
we artificially sort them in memory. This is OK because the local
filesystem impl is used only with small sizes and for tests.

Also added a function to run a block against all the test Storage
implementations. This function could be useful in other places, and we
could make it available outside the module.