Python can do distributed writes
It's quite hacky, we need better usability, but it works.

There is a new python test that uses Dask to do concurrent writers on
the same array.

Serialization works like this:

coordinator -> worker: we serialize a config dict, workers use it
to instantiate the Store

done worker -> coordinator: we serialize the bytes of the change set,
the coordinator uses that to do a distributed commit

Both paths can (and should) be improved, making objects picklable would
help, it's not trivial in pyo3 but doable.