Use AWS's S3 client instead of object_store
We are now using aws-sdk-s3 crate. object_store showed many issues and
performance problems while trying to use it with high concurrency.

We are still using object_store, but only for the in-memory and
local-filesystem `Storage` implementations, S3 (and minio) go through
aws-sdk-s3. So the dependency no longer includes the `aws` feature.

We did some changes to the virtual ref resolver. Now, we can use a
single instance of an S3 client, because the S3 client allows that,
unlike object_store that needs one client per bucket. This is a big
simplification of the code, at least while we have a single set of
credentials. We are no longer caching the object_store instances for
local filesystem refs.