Manifests shrink when there are fewer chunk references
In the previous algorithm we were copying all chunks from the previous
version to the new manifest (modulo changes in the current session).

Now, we only copy the chunks we need to copy. For example, after a
`clear` operation, there will be 0 references in the manifest.

Most of the commit is moving code around to make it accessible from
`distributed_flush`, that is, moving it from the `Repository` impl to
free functions. The only important change happens in
`distributed_flush`, where instead of starting from all the chunks in
the previous manifest, we use the `all_chunks` iterator.

A new test verifies manifests shrink when we delete chunks or whole
arrays.

Implements: #174