Node ids are random not sequential
We want this to avoid conflicts in (at least) two scenarios:

* distributed writers, where more than one creates a new node
* rebasing of commits, where multiple commits create nodes

This is not strictly needed, we could "renumber" the nodes as we merge
them, but it would increase complexity.

The downside of this change: slightly bigger type for nodes, slightly
more cloning.

WARNING: this is an on-disk format breaking change, Icechunk versions
with this change cannot read repositories written with previous
versions.