Allow overwriting existing groups and arrays
Closes #347

This bug manifests when groups or arrays were committed, and then are
being overwritten. Zarr will delete the existing nodes first and then
add them. This means the same node is in both the `new_*` and
`deleted_*` hashmaps, and so getting the node fails (it prioritizes
`deleted_groups` and returns KeyError)