Fetch chunks using concurrent requests (#553)
Only applies if the chunk is large enough. Same logic as with manifests.

This still doesn't apply to virtual chunks, only native. For no
reason, we can follow up with virtual chunks in a different PR.

This commit also fixes a bug in chunks with offset != 0 (wich cannot be
constructed in IC today)