Solve deadlock exposed by python threads
Many of our rust functions called from python use some form of blocking.
These functions need to release the Gil. Otherwise, if the lock that is to be
released to make progress is modified by a different thread, we have a
deadlock.

I don't know what the performance impact of this much calling to
`allow_threads` is, but I think it's either that or disallowing python threads.