Better defaulting mechanism for configuration (#613)
Now configurations can be merged, this means users can pass only the
configuration parameters they want to override, not the full object.
With this, users don't have to fetch the configuration to be able to
open a repository with some config changes.

Example:

```rust
Repository::open(
  ...,
  config=RepositoryConfig{inline_chunk_threshold_bytes=0, ..Default::default()})
```

Will override only `inline_chunk_threshold_bytes` and keep the rest of
the configuration as stored in the repo (or the default config if no
persistent configuration is found).

In addition, now we don't store the configuration for the repo on repo
creation, unless it's different from the default. This allows us to know
the user hasn't changed configuration, which will help with version
evolution.

To achieve all this:

* Now all configuration parameters across all structs are optional
* We introduce a `merge` function for configurations