Generate one manifest per array (#598)
* Generate one manifest per array

This is not an end state but an intermediate step. We won't release this
as a version, but once we have one manifest per array, it's easier to
group those arrays and pack them.

This commit also implements array extents tracking. So now, the
`ManifestRef` in the snapshot identifies, for each manifest, what part of
the array is contained there.

* Apply suggestions from code review

Co-authored-by: Deepak Cherian <dcherian@users.noreply.github.com>

* PR feedback

---------

Co-authored-by: Deepak Cherian <dcherian@users.noreply.github.com>