Uncache history
We had a cache in each snapshot with pointers to all the snapshot
ancestry. This makes expiration harder, because we need to invalidate
these caches.

We are replacing the cache by a single `parent_id` link. In the future,
we may optimize this for performance.

Notice: python is still turning the stream of parents into a python
list. This is now much worse because the new ancestry version is much
slower. We'll need to change this to use an AsyncGenerator.