Better error reporting and logs (#689)
* Better error reporting and logs

We use the tracing and tracing-subscriber crates for this.

Error types now include the tracing span where they were captured.
Converting from an error type to other must be done using `into` to
preserve the span.

Errors are much richer now, including the span where the error was
generated. To transfer this information to Python we use the exception
message.

If for some reason python users don't want to connect the tracing mechanism,
they can set the environment variable `ICECHUNK_NO_LOGS`. But this means
exceptions won't include the span information in their messages.

There is a new optional `feature` in the Rust library that adds log
enabling. `icechunk-python` uses the feature and enables logs in
`__init__.py`.

Code using Icechunk as a Rust library is at liberty of calling the new
Rust function `initialize_tracing` or not.

Logs are controlled by the `ICECHUNK_LOG` environment variable as
described [here](https://docs.rs/tracing-subscriber/latest/tracing_subscriber/filter/struct.EnvFilter.html#directives).

Logs are sent to stdout and are not integrated with python logging.

* Add FAQ entry

* update FAQ

* typo

---------

Co-authored-by: Deepak Cherian <deepak@earthmover.io>