Treat zarr metadata as a blob (mostly) (#749)
* Treat zarr metadata as a blob (mostly)

We were parsing too much of Zarr metadata. Icechunk currently is only
interested in the array size and chunk sizes. It may become interested
in the dimension names at some point. But still, we were parsing the
whole metadata, storing internally as parsed object and then formatting
it back to json.

We did this when the project started, imagining we may need more from
the metadata. For example, we thought we could need to incorporate the
codec pipeline in Icechunk.

With this patch, we now only extract the parts of the zarr metadata we
care about. And we preserve the original blob of metadata as is, in a
new user_data byte array. We return this blob in metadata requests.

If, in the future, we need more from the metadata, we can parse it and add
it to the storage.

Simpler and less code. It works with zarr extensions, it's more
resilient to zarr spec changes.

There is a price to this: we are no longer separating the user
attributes from the rest of the metadata. The only impact of this, is we
no longe can treat conflicts in user attributes separate from the rest
of the zarr metadata.

If we consider this important in the short term, we can add it back by
parsing more of the metadata blobs.

Also in this change:

- No more AttributeFile. We'll implement it when we need it
- Better snapshot serialization

[on-disk breaking change]

* Enable complex arrays in tests

* fix xarray test

---------

Co-authored-by: Deepak Cherian <deepak@earthmover.io>