Initial support for tigris consistency
In the default mode, we add the necessary headers to every request to
get Tigris to behave with high consistency.

We offer initial support for eventually consistent behavior, for
readonly datasets. This will require a bit more of work to improve
safety, because Tigris seems to very aggressively cache objects.