Avoid creating Rust Vec instances on the Python edge
Functions such as `zarr.Store.get` and `zarr.Store.set` were using
`Vec<u8>` as the interface mechanism. This proved to be very slow in
usages where there is high single threded concurrency.

We are now using
[PyBytes](https://docs.rs/pyo3-bytes/0.2.0/pyo3_bytes/index.html)
which uses unsafe Rust to avoid copying and allocation.

In tests, we get a bandwidth improvement between 10x and 20x when
copyng from and to local filesystem or in memory stores.