Speedup `list_prefix` and `getsize_prefix` when prefix is not empty
These two operations were very inefficient in repositories with a large
number of arrays. This PR implements pre-filtering of chunks, so we only
need to open manifests for the arrays that match the prefix.

Also in this PR: limit to 50 the number of preloaded array manifests.
Making this number configurable is an exercise left to a future PR.