Add `Session.flush` to allow anonymous snapshot creation (#1266)
Users can now choose to save their sessions to an anonymous
snapshot instead of committing, which updates a branch.

For this they use the `Session.flush` method. `flush` takes the same
arguments as `commit` (without the rebase-related arguments), and
returns the ID of the newly created snapshot.

The new snapshot is not pointed by any branches or tags.